# src/frontend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "VenueAxe.Web | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5280/"
    }
  ],
  "paths": {
...
    "/api/lanes/operations/{laneId}/start-session": {
      "post": {
        "tags": [
          "LaneOperations"
        ],
        "parameters": [
          {
            "name": "laneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveSessionSummaryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveSessionSummaryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveSessionSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/lanes/operations/{laneId}/active-session": {
      "get": {
        "tags": [
          "LaneOperations"
        ],
        "parameters": [
          {
            "name": "laneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveSessionSummaryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveSessionSummaryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveSessionSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/lanes/operations/{laneId}/throw": {
      "post": {
        "tags": [
          "LaneOperations"
        ],
        "parameters": [
          {
            "name": "laneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThrowInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ThrowInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ThrowInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              }
            }
          }
        }
      }
    },
    "/api/lanes/operations/{laneId}/undo": {
      "post": {
        "tags": [
          "LaneOperations"
        ],
        "parameters": [
          {
            "name": "laneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              }
            }
          }
        }
      }
    },
    "/api/lanes/operations/{laneId}/skip-turn": {
      "post": {
        "tags": [
          "LaneOperations"
        ],
        "parameters": [
          {
            "name": "laneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              }
            }
          }
        }
      }
    },
    "/api/lanes/operations/{laneId}/extend": {
      "post": {
        "tags": [
          "LaneOperations"
        ],
        "parameters": [
          {
            "name": "laneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendSessionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendSessionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lanes/operations/{laneId}/safety-stop": {
      "post": {
        "tags": [
          "LaneOperations"
        ],
        "parameters": [
          {
            "name": "laneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "Safety Briefing Required"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lanes/operations/{laneId}/rematch": {
      "post": {
        "tags": [
          "LaneOperations"
        ],
        "parameters": [
          {
            "name": "laneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              }
            }
          }
        }
      }
    },
    "/api/lanes/operations/{laneId}/switch-game": {
      "post": {
        "tags": [
          "LaneOperations"
        ],
        "parameters": [
          {
            "name": "laneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectGameRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectGameRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SelectGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateSnapshot"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lanes/operations/{laneId}/end-session": {
      "post": {
        "tags": [
          "LaneOperations"
        ],
        "parameters": [
          {
            "name": "laneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lanes/operations/{laneId}/substitute": {
      "post": {
        "tags": [
          "LaneOperations"
        ],
        "parameters": [
          {
            "name": "laneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubstitutePlayerRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubstitutePlayerRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubstitutePlayerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActiveSessionSummaryDto": {
        "required": [
          "sessionId",
          "sessionTitle",
          "startedAt",
          "expiresAt",
          "minutesRemaining",
          "activeRosterJson",
          "currentGame"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "sessionTitle": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "minutesRemaining": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "activeRosterJson": {
            "type": "string"
          },
          "currentGame": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GameStateSnapshot"
              }
            ]
          }
        }
      },
...
      "ExtendSessionRequest": {
        "required": [
          "extraMinutes"
        ],
        "type": "object",
        "properties": {
          "extraMinutes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "GameConfig": {
        "type": "object",
        "properties": {
          "totalRounds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "throwsPerRound": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "startingScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "allowClutchAnytime": {
            "type": "boolean"
          },
          "allowKillAnytime": {
            "type": "boolean"
          }
        }
      },
...
      "GamePlayer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatarColor": {
            "type": "string"
          },
          "score": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "throwsTaken": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "bullseyesHit": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "killsHit": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "clutchesHit": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "killsRemaining": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "killsCalled": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "streak": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "throwHistory": {
            "type": "array",
            "items": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        }
      },
      "GameStateSnapshot": {
        "type": "object",
        "properties": {
          "matchId": {
            "type": "string",
            "format": "uuid"
          },
          "gameTypeId": {
            "type": "string"
          },
          "gameName": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MatchStatus"
          },
          "currentRound": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalRounds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currentPlayerIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GamePlayer"
            }
          },
          "winnerPlayerId": {
            "type": [
              "null",
              "string"
            ]
          },
          "winnerName": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastThrow": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ThrowRecord"
              }
            ]
          },
          "allThrows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThrowRecord"
            }
          },
          "customDataJson": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "MatchStatus": {
        "type": "integer"
      },
...
      "SelectGameRequest": {
        "required": [
          "gameTypeId",
          "config"
        ],
        "type": "object",
        "properties": {
          "gameTypeId": {
            "type": "string"
          },
          "config": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GameConfig"
              }
            ]
          }
        }
      },
...
      "StartSessionRequest": {
        "required": [
          "sessionTitle",
          "durationMinutes",
          "initialRoster",
          "bookingId"
        ],
        "type": "object",
        "properties": {
          "sessionTitle": {
            "type": "string"
          },
          "durationMinutes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "initialRoster": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GamePlayer"
            }
          },
          "bookingId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "gameTypeId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SubstitutePlayerRequest": {
        "required": [
          "playerId",
          "newName"
        ],
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string"
          },
          "newName": {
            "type": "string"
          },
          "newAvatarColor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "TargetZone": {
        "type": "integer"
      },
...
      "ThrowInputDto": {
        "required": [
          "x",
          "y",
          "manualZone",
          "isClutchCalled"
        ],
        "type": "object",
        "properties": {
          "x": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "y": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "manualZone": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TargetZone"
              }
            ]
          },
          "isClutchCalled": {
            "type": "boolean"
          },
          "targetCellIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ThrowRecord": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string"
          },
          "playerName": {
            "type": "string"
          },
          "zone": {
            "$ref": "#/components/schemas/TargetZone"
          },
          "pointsAwarded": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "x": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "y": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "isKillCalled": {
            "type": "boolean"
          },
          "isClutchCalled": {
            "type": "boolean"
          },
          "isBullseye": {
            "type": "boolean"
          },
          "isKillshot": {
            "type": "boolean"
          },
          "thrownAt": {
            "type": "string",
            "format": "date-time"
          },
          "targetCellIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}